chore(test-data): create subdirectory to house sample data
