+++ models/channel/src/channel-draft/presets/index.ts
@@ -1,6 +1,4 @@
-import hub from './hub';
-import storeOther from './store-other';
-import storeUsa from './store-usa';
+import * as SampleDataFashion from './sample-data-fashion';
 import withInventorySupplyAndProductDistributionRoles from './with-inventory-supply-and-product-distribution-roles';
 import withInventorySupplyAndProductDistributionRolesNoAddress from './with-inventory-supply-and-product-distribution-roles-no-address';
 import withInventorySupplyRole from './with-inventory-supply-role';
@@ -8,9 +6,7 @@ import withInventorySupplyRoleNoAddress from './with-inventory-supply-role-no-ad
 import withProductDistributionRole from './with-product-distribution-role';
 
 const presets = {
-  hub,
-  storeOther,
-  storeUsa,
+  SampleDataFashion,
   withProductDistributionRole,
   withInventorySupplyRole,
   withInventorySupplyRoleNoAddress,
