+++ models/channel/src/channel-draft/presets/sample-data-fashion/index.ts
@@ -1,0 +1,11 @@
+import hub from './hub';
+import storeOther from './store-other';
+import storeUsa from './store-usa';
+
+const presets = {
+  storeUsa,
+  storeOther,
+  hub,
+};
+
+export default presets;
